clean up models
